server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
fer==22.5.1
numpy==1.24.3
Pillow==10.1.0
tensorflow==2.15.0
mtcnn==0.1.1
